Better advice context for the OSS Rules of Play rating
Part of #500
