+++ src/main/java/com/sap/oss/phosphor/fosstars/advice/oss/github/OssSecurityGithubAdvisor.java
@@ -1,17 +1,13 @@
 package com.sap.oss.phosphor.fosstars.advice.oss.github;
 
 import com.sap.oss.phosphor.fosstars.advice.CompositeAdvisor;
-import com.sap.oss.phosphor.fosstars.advice.oss.AbstractOssAdvisor.OssAdviceContextFactory;
 import com.sap.oss.phosphor.fosstars.advice.oss.ArtifactVersionAdvisor;
 import com.sap.oss.phosphor.fosstars.advice.oss.CodeqlAdvisor;
 import com.sap.oss.phosphor.fosstars.advice.oss.FindSecBugsAdvisor;
 import com.sap.oss.phosphor.fosstars.advice.oss.FuzzingAdvisor;
 import com.sap.oss.phosphor.fosstars.advice.oss.LgtmAdvisor;
 import com.sap.oss.phosphor.fosstars.advice.oss.MemorySafetyAdvisor;
-import com.sap.oss.phosphor.fosstars.advice.oss.OssAdviceContentYamlStorage.OssAdviceContext;
 import com.sap.oss.phosphor.fosstars.advice.oss.SecurityPolicyAdvisor;
-import com.sap.oss.phosphor.fosstars.model.subject.oss.GitHubProject;
-import java.util.Optional;
 
 /**
  * This is an advisor for open-source projects hosted on GitHub.
@@ -20,46 +16,16 @@ import java.util.Optional;
 public class OssSecurityGithubAdvisor extends CompositeAdvisor {
 
   /**
-   * A context factory that provides contexts for projects on GitHub.
-   */
-  private static final OssAdviceContextFactory CONTEXT_FACTORY = subject -> new OssAdviceContext() {
-
-    @Override
-    public Optional<String> lgtmProjectLink() {
-      if (subject instanceof GitHubProject) {
-        GitHubProject project = (GitHubProject) subject;
-        return Optional.of(
-            String.format("https://lgtm.com/projects/g/%s/%s",
-                project.organization().name(), project.name()));
-      }
-
-      return Optional.empty();
-    }
-
-    @Override
-    public Optional<String> suggestSecurityPolicyLink() {
-      if (subject instanceof GitHubProject) {
-        GitHubProject project = (GitHubProject) subject;
-        return Optional.of(
-            String.format("https://github.com/%s/%s/security/policy",
-                project.organization().name(), project.name()));
-      }
-
-      return Optional.empty();
-    }
-  };
-
-  /**
    * Create a new advisor.
    */
   public OssSecurityGithubAdvisor() {
     super(
-        new CodeqlAdvisor(CONTEXT_FACTORY),
-        new LgtmAdvisor(CONTEXT_FACTORY),
-        new SecurityPolicyAdvisor(CONTEXT_FACTORY),
-        new FindSecBugsAdvisor(CONTEXT_FACTORY),
-        new MemorySafetyAdvisor(CONTEXT_FACTORY),
-        new ArtifactVersionAdvisor(CONTEXT_FACTORY),
-        new FuzzingAdvisor(CONTEXT_FACTORY));
+        new CodeqlAdvisor(AdviceForGitHubContextFactory.INSTANCE),
+        new LgtmAdvisor(AdviceForGitHubContextFactory.INSTANCE),
+        new SecurityPolicyAdvisor(AdviceForGitHubContextFactory.INSTANCE),
+        new FindSecBugsAdvisor(AdviceForGitHubContextFactory.INSTANCE),
+        new MemorySafetyAdvisor(AdviceForGitHubContextFactory.INSTANCE),
+        new ArtifactVersionAdvisor(AdviceForGitHubContextFactory.INSTANCE),
+        new FuzzingAdvisor(AdviceForGitHubContextFactory.INSTANCE));
   }
 }
